src/user-services/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
datetime
python-jose[cryptography]
motor
